+++ packages/eyes-testcafe/src/sdk.js
@@ -38,14 +38,16 @@ class DecoratedEyesFactory extends sdk.EyesFactory {
     // set api wrapper
     const api = {
       async open(...args) {
-        let openArgs, config
+        let openArgs, config, driver
         if (args && args.length === 1 && TypeUtils.isObject(args[0]) && !spec.isDriver(args[0])) {
           const {t, appName, testName} = args[0]
           openArgs = [t, appName, testName]
           config = translateArgsToConfig({...applitoolsConfigJs, ...args[0]})
+          driver = t
         } else {
           openArgs = args
           config = translateArgsToConfig(applitoolsConfigJs)
+          driver = args[0]
         }
         failTestcafeOnDiff = config.failTestcafeOnDiff
         tapDirPath = config.tapDirPath
@@ -53,16 +55,31 @@ class DecoratedEyesFactory extends sdk.EyesFactory {
           eyesInstance.setConfiguration(config)
           eyesInstance.logger = new Logger(process.env.APPLITOOLS_SHOW_LOGS || config.getShowLogs())
         }
+        // driver health check, re: https://trello.com/c/xNCZNfPi
+        try {
+          await spec.executeScript(driver, () => {
+            return true
+          })
+        } catch (error) {
+          throw new Error(
+            `The browser is in an invalid state due to JS errors on the page that TestCafe is unable to handle. Try running the test with TestCafe's --skip-js-errors option enabled: https://devexpress.github.io/testcafe/documentation/reference/configuration-file.html#skipjserrors`,
+          )
+        }
         return await _open(...openArgs)
       },
       async checkWindow(args) {
         let preparedArgs
         if (TypeUtils.isObject(args)) {
-          if (args.tag && Object.keys(args).length === 1)
-            preparedArgs = {tag: args.tag, target: 'window', fully: true}
-          else if (args.window && !args.fully) preparedArgs = {...args, fully: true}
-          else preparedArgs = {...args}
-        } else preparedArgs = {tag: args, target: 'window', fully: true}
+          preparedArgs = {...args}
+          if (!args.hasOwnProperty('target')) {
+            preparedArgs.target = 'window'
+          }
+          if (preparedArgs.target === 'window' && !args.hasOwnProperty('fully')) {
+            preparedArgs.fully = true
+          }
+        } else {
+          preparedArgs = {tag: args, target: 'window', fully: true}
+        }
 
         await _check(translateArgsToCheckSettings(preparedArgs))
       },