+++ packages/screenshoter/src/screenshoter.js
@@ -2,27 +2,31 @@ const utils = require('@applitools/utils')
 const makeScroller = require('./scroller')
 const takeStitchedScreenshot = require('./takeStitchedScreenshot')
 const takeViewportScreenshot = require('./takeViewportScreenshot')
+const scrollIntoViewport = require('./scrollIntoViewport')
 
 async function screenshoter({
   logger,
   driver,
-  context,
+  frames = [],
   target,
-  isFully,
+  fully,
   hideScrollbars,
   hideCaret,
   scrollingMode,
   overlap,
   wait,
-  rotate,
-  crop,
-  scale,
+  dom,
+  stabilization,
   debug,
+  takeDomCapture,
 }) {
   const originalContext = driver.currentContext
   const defaultScroller = makeScroller({logger, scrollingMode})
 
-  const targetContext = context ? await originalContext.context(context) : originalContext
+  const targetContext =
+    frames.length > 0
+      ? await originalContext.context(frames.reduce((parent, frame) => ({...frame, parent}), null))
+      : originalContext
   const scrollingStates = []
   for (const nextContext of targetContext.path) {
     const scrollingElement = await nextContext.getScrollRootElement()
@@ -33,22 +37,47 @@ async function screenshoter({
 
   const activeElement = hideCaret && !driver.isNative ? await targetContext.blurElement() : null
 
-  const area = await getTargetArea({logger, context: targetContext, target, isFully, scrollingMode})
+  const {context, scroller, region} = await getTargetArea({
+    logger,
+    context: targetContext,
+    target,
+    fully,
+    scrollingMode,
+  })
 
-  // temporary solution
-  if (isFully) {
-    await area.context.execute(
-      'arguments[0].setAttribute("data-applitools-scroll", "true")',
-      area.scroller.element,
-    )
-  }
-  // ---
+  const scrollerState = await scroller.getState()
+
+  await scrollIntoViewport({logger, context, scroller, region})
 
   try {
-    return isFully
-      ? await takeStitchedScreenshot({...area, logger, rotate, crop, scale, wait, overlap, debug})
-      : await takeViewportScreenshot({...area, logger, rotate, crop, scale, debug})
+    const screenshot = fully
+      ? await takeStitchedScreenshot({
+          logger,
+          context,
+          scroller,
+          region,
+          overlap,
+          wait,
+          stabilization,
+          debug,
+        })
+      : await takeViewportScreenshot({logger, context, region, wait, stabilization, debug})
+
+    if (!dom) return screenshot
+
+    // temporary solution
+    if (fully) {
+      await context.execute(
+        'arguments[0].setAttribute("data-applitools-scroll", "true")',
+        scroller.element,
+      )
+    }
+
+    return {...screenshot, dom: await takeDomCapture()}
+    // ---
   } finally {
+    await scroller.restoreState(scrollerState)
+
     if (hideCaret && activeElement) await targetContext.focusElement(activeElement)
 
     for (const prevContext of targetContext.path.reverse()) {
@@ -62,7 +91,7 @@ async function screenshoter({
   }
 }
 
-async function getTargetArea({logger, context, target, isFully, scrollingMode}) {
+async function getTargetArea({logger, context, target, fully, scrollingMode}) {
   if (target) {
     if (utils.types.has(target, ['x', 'y', 'width', 'height'])) {
       const scrollingElement = await context.getScrollRootElement()
@@ -75,7 +104,7 @@ async function getTargetArea({logger, context, target, isFully, scrollingMode}) 
       const element = await context.element(target)
       if (!element) throw new Error('Element not found!')
 
-      if (isFully) {
+      if (fully) {
         const isScrollable = await element.isScrollable()
         const scrollingElement = isScrollable ? element : await context.getScrollRootElement()
         return {
@@ -96,7 +125,7 @@ async function getTargetArea({logger, context, target, isFully, scrollingMode}) 
         }
       }
     }
-  } else if (!context.isMain && !isFully) {
+  } else if (!context.isMain && !fully) {
     const scrollingElement = await context.parent.getScrollRootElement()
     const element = await context.getFrameElement()
     return {