+++ packages/screenshoter/src/takeViewportScreenshot.js
@@ -1,25 +1,12 @@
 const utils = require('@applitools/utils')
 const makeTakeScreenshot = require('./takeScreenshot')
 const saveScreenshot = require('./saveScreenshot')
-const scrollIntoViewport = require('./scrollIntoViewport')
 
-async function takeViewportScreenshot({
-  logger,
-  context,
-  scroller,
-  region,
-  rotate,
-  crop,
-  scale,
-  wait,
-  debug = {},
-}) {
+async function takeViewportScreenshot({logger, context, region, wait, stabilization, debug = {}}) {
   logger.verbose('Taking image of...')
 
-  await scrollIntoViewport({logger, context, scroller, region})
-
   const driver = context.driver
-  const takeScreenshot = makeTakeScreenshot({logger, driver, rotate, crop, scale, debug})
+  const takeScreenshot = makeTakeScreenshot({logger, driver, stabilization, debug})
 
   const image = await takeScreenshot()