+++ packages/sdk-coverage-tests/src/generate/save.js
@@ -18,7 +18,7 @@ async function createTestMetaData(tests, {metaDir = '', pascalizeTests = true} =
   fs.mkdirSync(targetDirectory, {recursive: true})
 
   const meta = tests.reduce((meta, test) => {
-    const data = {isGeneric: true, name: test.group}
+    const data = {isGeneric: true, name: test.group, skip: test.skip, skipEmit: test.skipEmit}
     if (test.config) {
       if (test.config.stitchMode) data.executionMode = test.config.stitchMode.toLowerCase()
       else if (test.vg) data.executionMode = 'visualgrid'