+++ packages/sdk-release-kit/src/cli/index.js
@@ -1,30 +1,7 @@
 #!/usr/bin/env node
 
-const args = require('yargs')
-  .command(
-    ['preversion', 'release-pre-check', 'pre-version'],
-    'Run all verification checks pre-release',
-  )
-  .command(['version'], 'Supportive steps to version a package')
-  .command(['postversion, post-version'], 'Supportive steps to after a package has been versioned')
-  .command(['lint', 'l'], 'Static code analysis ftw')
-  .command(['verify-changelog', 'vch'], 'Verify changelog has unreleased entries')
-  .command(
-    ['verify-commits', 'vco'],
-    'Verify no unreleased changes for internal dependencies exist',
-  )
-  .command(['verify-versions', 'vv'], 'Verify consistent versions in relevant packages')
-  .command(['verify-installed-versions', 'viv'], 'Verify correct dependencies are installable')
-  .command(['ls-dry-run', 'ls'], 'Display dependencies from a verify-installed-versions run')
-  .command(['update-changelog', 'uc'], 'Create release entry in the changelog')
-  .command(
-    ['send-release-notification', 'hello-world'],
-    'Send a notification that a has been released',
-  )
-  .demandCommand()
-  .help().argv
+const yargs = require('yargs')
 const chalk = require('chalk')
-const cwd = process.cwd()
 const path = require('path')
 const fs = require('fs')
 const {verifyChangelog, writeReleaseEntryToChangelog} = require('../changelog')
@@ -33,133 +10,191 @@ const {lint} = require('../lint')
 const sendReleaseNotification = require('../send-report')
 const {createDotFolder} = require('../setup')
 const {verifyCommits, verifyInstalledVersions, verifyVersions} = require('../versions')
-const {gitAdd, gitCommit, gitPushWithTags, isChanged} = require('../git')
+const {gitAdd, gitCommit, gitPushWithTags, isChanged, gitStatus} = require('../git')
 const {yarnInstall, yarnUpgrade, verifyUnfixedDeps} = require('../yarn')
-const {should} = require('chai')
 
-const command = args._[0]
-
-;(async () => {
-  try {
-    switch (command) {
-      case 'lint':
-      case 'l':
-        return await lint(cwd)
-      case 'ls-dry-run':
-      case 'ls':
-        return lsDryRun()
-      case 'postversion':
-      case 'post-version':
-        await gitPushWithTags()
-        if (!args['skip-release-notification']) {
-          await sendReleaseNotification(cwd, args.recipient)
-        }
-        return
-      case 'preversion':
-      case 'pre-version':
-      case 'release-pre-check':
-        if (!args.skipDeps) {
-          console.log('[bongo preversion] yarn install')
-          await yarnInstall()
-        }
-        console.log('[bongo preversion] lint')
-        await lint(cwd)
-        console.log('[bongo preversion] verify changelog')
-        await verifyChangelog(cwd)
-        console.log('[bongo preversion] verify unfixed dependencies')
-        verifyUnfixedDeps(cwd)
-        if (!args.skipVerifyVersions) {
-          console.log('[bongo preversion] verify commits')
-          await verifyCommits({pkgPath: cwd}).catch(err => console.log(err.message))
-        }
-        try {
-          console.log('[bongo preversion] verify versions')
-          verifyVersions({pkgPath: cwd})
-        } catch (err) {
-          console.log(chalk.yellow(err.message))
-        }
-        if (!args.skipVerifyInstalledVersions) {
-          console.log('[bongo preversion] verify installed versions')
-          createDotFolder(cwd)
-          await packInstall(cwd)
-          await verifyInstalledVersions({
-            pkgPath: cwd,
-            installedDirectory: path.join('.bongo', 'dry-run'),
-          })
-        }
-        await commitFiles(!!args.skipCommit)
-        console.log('[bongo preversion] done!')
-        return
-      case 'send-release-notification':
-      case 'hello-world':
-        return await sendReleaseNotification(cwd, args.recipient)
-      case 'update-changelog':
-      case 'uc':
-        return writeReleaseEntryToChangelog(cwd)
-      case 'verify-changelog':
-      case 'vch':
-        return await verifyChangelog(cwd)
-      case 'verify-commits':
-      case 'vco':
-        return await verifyCommits({pkgPath: cwd})
-      case 'verify-installed-versions':
-      case 'viv':
+yargs
+  .config({cwd: process.cwd()})
+  .command(
+    ['preversion', 'release-pre-check', 'pre-version'],
+    'Run all verification checks pre-release',
+    {
+      skipVerifyInstalledVersions: {alias: 'sviv', type: 'boolean'},
+      skipVerifyVersions: {alias: 'svv', type: 'boolean'},
+      skipDeps: {alias: 'sd', type: 'boolean'},
+      skipCommit: {alias: 'sc', type: 'boolean', default: false},
+    },
+    async args => {
+      const {cwd} = args
+      if (!args.skipDeps) {
+        console.log('[bongo preversion] yarn install')
+        await yarnInstall()
+      }
+      console.log('[bongo preversion] lint')
+      await lint(cwd)
+      console.log('[bongo preversion] verify changelog')
+      verifyChangelog(cwd)
+      console.log('[bongo preversion] verify unfixed dependencies')
+      verifyUnfixedDeps(cwd)
+      if (!args.skipVerifyVersions) {
+        console.log('[bongo preversion] verify commits')
+        await verifyCommits({pkgPath: cwd}).catch(err => console.log(err.message))
+      }
+      try {
+        console.log('[bongo preversion] verify versions')
+        verifyVersions({pkgPath: cwd})
+      } catch (err) {
+        console.log(chalk.yellow(err.message))
+      }
+      if (!args.skipVerifyInstalledVersions) {
+        console.log('[bongo preversion] verify installed versions')
         createDotFolder(cwd)
         await packInstall(cwd)
-        return await verifyInstalledVersions({
+        await verifyInstalledVersions({
           pkgPath: cwd,
           installedDirectory: path.join('.bongo', 'dry-run'),
         })
-      case 'verify-versions':
-      case 'vv':
-        try {
-          verifyVersions({pkgPath: cwd})
-        } catch (err) {
-          console.log(chalk.yellow(err.message))
+      }
+      await commitFiles(args)
+      console.log('[bongo preversion] done!')
+    },
+  )
+  .command(['version'], 'Supportive steps to version a package', {}, async ({cwd}) => {
+    writeReleaseEntryToChangelog(cwd)
+    await gitAdd('CHANGELOG.md')
+  })
+  .command(
+    ['postversion, post-version'],
+    'Supportive steps to after a package has been versioned',
+    {
+      recipient: {alias: 'r', type: 'string'},
+      skipReleaseNotification: {alias: 'sr', type: 'boolean'},
+    },
+    async args => {
+      try {
+        console.log('[bongo postversion] pushing with tags')
+        await gitPushWithTags()
+        if (args.skipReleaseNotification) {
+          console.log('[bongo postversion] skipping release notification')
+        } else if (!args.skipReleaseNotification) {
+          console.log('[bongo postversion] sending release notification')
+          await sendReleaseNotification(args.cwd, args.recipient)
+          console.log('[bongo postversion] release notification sent')
         }
-        return
-      case 'version':
-        writeReleaseEntryToChangelog(cwd)
-        return await gitAdd('CHANGELOG.md')
-      case 'deps':
-        await deps()
-        return await commitFiles(!!args.skipCommit)
-      default:
-        throw new Error('Invalid option provided')
+        console.log('[bongo postversion] done!')
+      } catch (err) {
+        console.log(chalk.yellow(err.message))
+      }
+    },
+  )
+  .command(['lint', 'l'], 'Static code analysis ftw', {}, async ({cwd}) => await lint(cwd))
+  .command(['verify-changelog', 'vch'], 'Verify changelog has unreleased entries', {}, ({cwd}) =>
+    verifyChangelog(cwd),
+  )
+  .command(
+    ['verify-commits', 'vco'],
+    'Verify no unreleased changes for internal dependencies exist',
+    {},
+    async ({cwd}) => await verifyCommits({pkgPath: cwd}),
+  )
+  .command(
+    ['verify-versions', 'vv'],
+    'Verify consistent versions in relevant packages',
+    {},
+    async ({cwd}) => {
+      try {
+        verifyVersions({pkgPath: cwd})
+      } catch (err) {
+        console.log(chalk.yellow(err.message))
+      }
+    },
+  )
+  .command(
+    ['verify-installed-versions', 'viv'],
+    'Verify correct dependencies are installable',
+    {},
+    async ({cwd}) => {
+      createDotFolder(cwd)
+      await packInstall(cwd)
+      await verifyInstalledVersions({
+        pkgPath: cwd,
+        installedDirectory: path.join('.bongo', 'dry-run'),
+      })
+    },
+  )
+  .command(
+    ['ls-dry-run', 'ls'],
+    'Display dependencies from a verify-installed-versions run',
+    {},
+    () => lsDryRun(),
+  )
+  .command(['update-changelog', 'uc'], 'Create release entry in the changelog', {}, ({cwd}) =>
+    writeReleaseEntryToChangelog(cwd),
+  )
+  .command(
+    ['send-release-notification', 'hello-world'],
+    'Send a notification that an sdk has been released',
+    {recipient: {alias: 'r', type: 'string'}},
+    async args => await sendReleaseNotification(args.cwd, args.recipient),
+  )
+  .command(
+    ['deps', 'd'],
+    'update internal deps',
+    {
+      skipCommit: {type: 'boolean', default: false},
+      upgradeAll: {type: 'boolean', default: false},
+    },
+    async args => {
+      console.log('[bongo deps] running...')
+      await deps(args)
+      console.log('[bongo deps] updated deps. now committing...')
+      await commitFiles(args)
+      console.log('[bongo deps] done')
+    },
+  )
+  .demandCommand(1, 'exit')
+  .fail((msg, error, args) => {
+    if (msg === 'exit') {
+      return args.showHelp()
     }
-  } catch (error) {
-    if (args.verbose) {
+    const command = process.argv[2]
+    if (process.argv.includes('--verbose')) {
       console.log(error)
     } else {
       console.log(chalk.red(error.message))
       console.log(`run "npx bongo ${command} --verbose" to see stack trace`)
     }
     process.exit(1)
-  }
-})()
+  })
+  .wrap(yargs.terminalWidth())
+  .help().argv
 
-async function deps() {
+async function deps({cwd, upgradeAll}) {
   verifyUnfixedDeps(cwd)
   await yarnUpgrade({
     folder: cwd,
-    upgradeAll: args.upgradeAll,
+    upgradeAll,
   })
 }
 
-async function commitFiles(skipCommit = false) {
+async function commitFiles({cwd, skipCommit}) {
   if (!skipCommit) {
+    console.log('[bongo] commit files running...\n', (await gitStatus()).stdout)
     const files = ['package.json', 'CHANGELOG.md', 'yarn.lock']
     for (const file of files) {
       // git add fails when trying to add files that weren't changed
       if (await isChanged(file)) {
+        console.log(`[bongo] git add changed file: ${file}`)
         await gitAdd(file)
       }
     }
 
     // git commit fails when trying to commit files that weren't changed
+    console.log(`[bongo] committing changed files:\n${(await gitStatus()).stdout}`)
     if (await isChanged(...files)) {
       const pkgName = JSON.parse(fs.readFileSync(path.resolve(cwd, 'package.json'))).name
       await gitCommit(`[auto commit] ${pkgName}: upgrade deps`)
+      console.log(`[bongo] actually committed files`)
     }
   }
 }