+++ packages/driver/CHANGELOG.md
@@ -3,6 +3,16 @@
 ## Unreleased
 
 
+## 1.0.4 - 2021/1/27
+
+- no changes
+
+## 1.0.3 - 2021/1/27
+
+- chore: add husky
+- updated to @applitools/snippets@2.1.1 (from 2.1.0)
+- updated to @applitools/utils@1.0.1 (from 1.0.0)
+
 ## 1.0.2 - 2020/12/29
 
 - add `setTransforms` method to element