+++ packages/eyes-nightwatch/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+
+## 1.5.2 - 2021/1/29
+
+- chore: add husky
+- updated to @applitools/eyes-sdk-core@12.14.2 (from 12.12.2)
+- updated to @applitools/visual-grid-client@15.5.14 (from 15.5.5)
+
 ## 1.5.1 - 2021/1/12
 
 - updated to @applitools/eyes-sdk-core@12.12.2 (from 12.10.0)