+++ packages/eyes-playwright/CHANGELOG.md
@@ -3,6 +3,12 @@
 ## Unreleased
 
 
+## 1.5.1 - 2021/1/29
+
+- chore: add husky
+- updated to @applitools/eyes-sdk-core@12.14.2 (from 12.10.0)
+- updated to @applitools/visual-grid-client@15.5.14 (from 15.4.0)
+- updated to @applitools/visual-grid-client@15.5.14 (from 15.4.0)
 ## 1.5.0 - 2020/12/18
 
 - updated to @applitools/eyes-sdk-core@12.10.0 (from 12.9.2)