+++ packages/eyes-protractor/CHANGELOG.md
@@ -3,6 +3,12 @@
 ## Unreleased
 
 
+## 1.13.2 - 2021/1/29
+
+- chore: add husky
+- updated to @applitools/eyes-sdk-core@12.14.2 (from 12.12.2)
+- updated to @applitools/visual-grid-client@15.5.14 (from 15.5.5)
+
 ## 1.13.1 - 2021/1/12
 
 - updated to @applitools/eyes-sdk-core@12.12.2 (from 12.10.0)