+++ packages/eyes-selenium/CHANGELOG.md
@@ -3,6 +3,19 @@
 ## Unreleased
 
 
+## 4.52.5 - 2021/2/1
+
+- updated to @applitools/eyes-sdk-core@12.14.6 (from 12.14.1)
+- updated to @applitools/visual-grid-client@15.5.19 (from 15.5.13)
+- updated to @applitools/eyes-sdk-core@12.14.7 (from 12.14.6)
+- updated to @applitools/visual-grid-client@15.5.20 (from 15.5.19)
+
+## 4.52.4 - 2021/1/27
+
+- chore: add husky
+- updated to @applitools/eyes-sdk-core@12.14.1 (from 12.13.2)
+- updated to @applitools/visual-grid-client@15.5.13 (from 15.5.8)
+
 ## 4.52.3 - 2021/1/15
 
 - updated to @applitools/eyes-sdk-core@12.13.2 (from 12.12.2)