+++ packages/eyes-storybook/CHANGELOG.md
@@ -3,6 +3,20 @@
 ## Unreleased
 
 
+## 3.16.5 - 2021/1/31
+
+- updated to @applitools/dom-snapshot@4.4.10 (from 4.4.9)
+- updated to @applitools/eyes-sdk-core@12.14.6 (from 12.14.5)
+- updated to @applitools/visual-grid-client@15.5.19 (from 15.5.17)
+
+## 3.16.4 - 2021/1/31
+
+- chore: add husky
+- updated to @applitools/dom-snapshot@4.4.9 (from 4.4.7)
+- updated to @applitools/eyes-puppeteer@1.4.3 (from 1.4.2)
+- updated to @applitools/eyes-sdk-core@12.14.5 (from 12.13.4)
+- updated to @applitools/visual-grid-client@15.5.17 (from 15.5.10)
+
 ## 3.16.3 - 2021/1/15
 
 - updated to @applitools/dom-snapshot@4.4.7 (from 4.4.5)