+++ packages/eyes-testcafe/CHANGELOG.md
@@ -3,6 +3,21 @@
 ## Unreleased
 
 
+## 1.12.5 - 2021/2/2
+
+- fix default checkWindow behavior- take full page when no target is specified
+- set minimum supported Node version to be consistent with other SDKs (8.9)
+- updated to @applitools/eyes-sdk-core@12.14.7 (from 12.14.2)
+- updated to @applitools/visual-grid-client@15.5.20 (from 15.5.13)
+
+## 1.12.4 - 2021/1/27
+
+- chore: add husky
+- update peer dependency version of testcafe to include earlier versions (e.g., 1.7.x)
+- fix issue where the browser is in an invalid state but eyes.open tries to proceed ([Trello](https://trello.com/c/xNCZNfPi))
+- updated to @applitools/eyes-sdk-core@12.14.2 (from 12.13.0)
+- updated to @applitools/visual-grid-client@15.5.13 (from 15.5.6)
+
 ## 1.12.3 - 2021/1/13
 
 - fix for fetching font-face resources from stylesheets [Trello](https://trello.com/c/DwmxtRoR)