+++ packages/eyes-webdriverio-5-service/CHANGELOG.md
@@ -3,6 +3,11 @@
 ## Unreleased
 
 
+## 1.11.2 - 2021/1/29
+
+- chore: add husky
+- updated to @applitools/eyes-webdriverio@5.29.4 (from 5.29.1)
+
 ## 1.11.1 - 2021/1/12
 
 - updated to @applitools/eyes-webdriverio@5.29.1 (from 5.29.0)