+++ packages/sdk-fake-eyes-server/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- chore: add husky 
 
 ## 2.1.1 - 2020/11/24