+++ packages/sdk-release-kit/CHANGELOG.md
@@ -3,6 +3,19 @@
 ## Unreleased
 
 
+## 0.12.2 - 2021/2/2
+
+- add logging to postversion
+- updated to @applitools/sdk-shared@0.8.6 (from 0.8.5)
+
+## 0.12.1 - 2021/1/27
+
+- chore: add husky
+
+## 0.12.0 - 2021/1/20
+
+- yargs as god intended
+
 ## 0.11.1 - 2021/1/12
 
 - updated to @applitools/sdk-shared@0.8.5 (from 0.8.0)