+++ packages/sdk-shared/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- chore: add husky
 
 ## 0.8.5 - 2021/1/11