+++ packages/snippets/CHANGELOG.md
@@ -4,6 +4,11 @@
 ## Unreleased
 
 
+## 2.1.1 - 2021/1/26
+
+- handle translated html elements during content size extracting
+- chore: add husky
+
 ## 2.1.0 - 2020/10/23
 
 - fix `blurElement`: use default value if element passed as `null`