+++ packages/utils/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+
+## 1.0.1 - 2021/1/27
+
+- chore: add husky
+
 ## 1.0.0 - 2020/12/1
 
 - Provide general utils, geometry utils, types utils and argument guard written with ts